scripts/requirements.txt
scholarly
beautifulsoup4
requests
pdfplumber
PyYAML
rich
